pyproject.toml
[tool.poetry]
name = "fastapi_restaurant"
version = "0.1.0"
description = ""
authors = ["Abdelrahman A. Elsharawy <aelsharway@stcs.com.sa>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = {extras = ["uvicorn"], version = "^0.85.0"}
uvicorn = "^0.13.4"
gunicorn = "^20.1.0"
alembic = "^1.6.2"
psycopg2 = "^2.8.6"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {extras = ["dotenv"], version = "^1.8.2"}
python-jose = {extras = ["cryptography"], version = "^3.2.0"}
sqladmin = "^0.6.0"
fastapi-pagination = "^0.10.0"
aioredis = "^2.0.1"
fastapi-redis-cache = "^0.2.5"
pytz = "^2022.2.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-cov = "^2.10.1"
requests = "^2.25.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
